+++ tests/client/settings.py
@@ -19,7 +19,7 @@
 # However, if you have executed another commercial license agreement
 # with Crate these terms will supersede the license and you may use the
 # software solely pursuant to the terms of the relevant commercial agreement.
-from __future__ import absolute_import
+
 
 from pathlib import Path